fix(linter): move no-use-before-define to eslint plugin (#19438)
fixes #19437
fixes https://github.com/oxc-project/oxc/issues/19542
